Revert "Implement Timer-A emulation (delay-mode only)."
This reverts commit dc693d130c76cc945269c64c1467f9c6eb45e670.

+++ src/Unix/host_clock_unix.cpp
@@ -25,10 +25,10 @@
 
 /*--- Private stuff ---*/
 
-uint32 HostClockUnix::getCurTime(void)
+float HostClockUnix::getCurTime(void)
 {
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
 
-	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
+	return ((tv.tv_sec * 1000.0) + (tv.tv_usec / 1000.0));
 }
